UPSTREAM:PROTOCOL:hostapd:Handle address translation for FT  in assoc response
When the AP sends a response for a reassoc/assoc request in Fast
Transition (FT), the link address is incorrectly overwritten by
the station's MLD address if the station is found in the FT roaming
list. This prevents the station from receiving the assoc response.
Remove the address reassignment so that the reassoc/assoc response
is sent to the station's link address.

Fixes: I3e66cbd27c1(UPSTREAM:PROTOCOL:hostapd: Support MLD STA
reconnection on same band)
Patch-dependency: s021-UPSTREAM-PROTOCOL-hostapd-Support-MLD-STA-reconnecti.patch

Change-Id: I9a27ef11cbd6b56a84f13c061b1bf482f020c642
Signed-off-by: Sowmiya Sree Elavalagan <ssreeela@qti.qualcomm.com>