UPSTREAM:PROTOCOL:HOSTAPD: Remove STA if parsing  peer_commit_element_ecc fails in SAE commit
When an AP sends an SAE confirm message, it starts a retransmit timer. In some cases,
the STA does not respond to the confirm message but instead sends an SAE commit
message again. This behavior can occur if the STA did not receive the SAE confirm
message from the AP and restarted the SAE session.

While parsing the SAE commit message, if the AP fails to parse peer_commit_element_ecc,
it currently responds to the STA with an unspecified failure status. However,
peer_commit_element_ecc remains set to NULL. Later, when the retransmit timer triggers
and attempts to send an SAE confirm message, the missing peer_commit_element_ecc
causes a crash.

To prevent this, remove the STA entry when parsing peer_commit_element_ecc
fails during SAE commit processing.

Patch-work: None
Patch-dependency: None

Change-Id: I337f53649b3c1900cc580a304871bccfa714fd41
Signed-off-by: Arunpandi Kannan <arunpand@qti.qualcomm.com>