QSDK:PROTOCOL:Hostapd: Remove BCCA IE with zero color switch countdown from Beacon
Currently, in the color change command, the beacon template applied
after the color switch countdown reaches 1 includes a BCCA IE with
the color switch countdown set to 0. As per 802.11 specification,
a color switch countdown of zero is reserved in the BCCA IE and is
not required in BSS color change procedure. Remove the BCCA IE
from the beacon template used when the color switch countdown reaches 1.

The corresponding driver logic to include the BCCA IE with a
color switch countdown of zero is removed through a separate gerrit.
https://review-android.quicinc.com/c/oss/src/mac80211/wlan-open/+/6438509

Patch-dependency: none

Change-Id: Id1d29c8cfe0e9e844b29f1a26de7183387fce48b
Signed-off-by: Suraj P Kizhakkethil <suraj.kizhakkethil@oss.qualcomm.com>