QSDK:PROTOCOL: Makefile: Enable External-Plugin
Allow external application to plugin-in library statically to hostapd
there-by registering a plugin object with hostapd. Once "plugged-in",
hostapd-interface layer will use the plugin object to call into the
external library, and also initialize some of the inbound handlers which
external library will use to call in to hostapd.

Patch-dependency: None

Change-Id: Ibdb409b9c4adf9d65f7fe00b0b02a97b1c563f0d
Signed-off-by: Adil Saeed Musthafa <adilm@qti.qualcomm.com>